refactor(rpc): isolate reply parsing module
